SferaAI_2/requirements.txt
livekit-agents
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
fastembed==0.5.1
numpy>=1.24.0
ddgs
langchain_community
langchain_experimental
langchain-google-genai
google-generativeai
langgraph
aiohttp
python-dotenv
pytest
pytest-asyncio
flake8
tinydb
youtube-search-python
redis>=5.0.0
sentry-sdk>=1.40.0